requirements.txt
flask==3.0.3
openai==2.29.0
python-dotenv==1.2.1
